(少選課人數變更、必選修輸出)

+++ rear/main.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # coding=utf-8
 # -*- coding: UTF-8 -*-
-#uvicorn main:app --reload
+#欲啟用後端請輸入:uvicorn main:app --reload
 
 from xmlrpc.client import boolean
 from fastapi import FastAPI
@@ -36,7 +36,6 @@ c = pymysql.connect(**db_settings)
 cursor = c.cursor(pymysql.cursors.DictCursor)
 cursor2 = c.cursor()
 
-
 def checkUID(UID):  #確認帳號是否存在
     command = f"SELECT COUNT(*) FROM `student` WHERE `StudentID` = \"{UID}\""
     cursor2.execute(command)
@@ -56,18 +55,22 @@ def checkselected(UID):  #確認是否有選課
     else:
         return True
     
+    
 
 def checkcredit(CID,data): #確認學分是否足夠
     totalcredit = 0
-    # print("checkcredit CID type and data type",CID[0],data[0][0],type(CID),type(data))
-    for i in range(len(data)):
-        transtemp = data[i][0]
-        if(int(transtemp) != int(CID)):
-            totalcredit += data[i][8]
-    if(totalcredit>8):
-        return True
+    if(len(data)>0):
+        for i in range(len(data)):
+            transtemp = data[i][0]
+            if(int(transtemp) != int(CID)):
+                totalcredit += data[i][8]
+        if(totalcredit>8):
+            return True
+        else:
+            return False
     else:
         return False
+
 
 
 def delra(ra,S_CID):
@@ -107,14 +110,12 @@ def opcourse_A():  # 驗證選課人數 並去除人數已滿之課程
     cursor2.execute(command)
     r1 = cursor2.fetchall()
     r = list(r1)
-    print("RA剩餘:",len(r))
     return r
 
 
 def opcourse_B(ra,S_CIDs):  # 去除已選課程
     for i in range(len(S_CIDs)):
         rb = delra(ra,S_CIDs[i])
-    print("RB剩餘:",len(rb))
     return rb
 
     
@@ -130,7 +131,6 @@ def opcourse_C(rb,S_CIDs):  # 去除學分不合條件課程
     lastcredit = 30-totalcredit
     for i in range(len(rb)):
         rc = delrb(rb,lastcredit)
-    print("RC剩餘:",len(rb))
     return rc
 
 
@@ -145,7 +145,6 @@ def opcourse_D(rc,S_CIDs):  # 去除相同時間之課程
     for i in range(len(rc)):
         for j in range(len(Schedule)):
             rd = delrc(rc,Schedule[j])
-    print("RD剩餘:",len(rd))
     return rd
 
 
@@ -160,7 +159,6 @@ def opcourse_E(rd,S_CIDs):  # 去除相同課名之課程
     for i in range(len(rd)):
         for j in range(len(CourseName)):
             re = delrd(rd,CourseName[j])
-    print("RE剩餘:",len(re))
     return re
 
 
@@ -171,21 +169,20 @@ def opcourse(UID): #可選課程篩選
         command = f"SELECT CourseID FROM `selectedcourse` where StudentID = \"{UID}\";"
         cursor2.execute(command)
         rr = cursor2.fetchall()
-
         for i in range(len(rr)):
             S_CIDs.append(rr[i][0])
         rb = opcourse_B(ra,S_CIDs)
-        if(rb==None):
-            return None
+        if(len(rb)==0):
+            return []
         rc = opcourse_C(rb,S_CIDs)
-        if(rc==None):
-            return None
+        if(len(rc)==0):
+            return []
         rd = opcourse_D(rc,S_CIDs)
-        if(rd==None):
-            return None
+        if(len(rd)==0):
+            return []
         re = opcourse_E(rd,S_CIDs)
-        if(re==None):
-            return None
+        if(len(re)==0):
+            return []
         return re
         
     else:
@@ -193,12 +190,45 @@ def opcourse(UID): #可選課程篩選
 
 
 def checkCID(CID,data): #確認課號是否存在
-    # print("CID type and data type",CID[0],data[0][0],type(CID),type(data))
-    for i in range(len(data)):
-        transtemp = data[i][0]
-        if(int(CID) == int(transtemp)):
-            return True
-    return False  
+    if(len(data)>0):
+        for i in range(len(data)):
+            transtemp = data[i][0]
+            if(int(CID) == int(transtemp)):
+                return True
+        return False
+    else:
+        return False 
+
+
+def data_return(UID): #資料回傳
+    global data_a, data_b
+    data_a = []
+    data_b = []
+    if(checkselected(UID)):
+        command = f"SELECT CourseID FROM `selectedcourse` where StudentID = \"{UID}\";"
+        cursor2.execute(command)
+        S_CID = cursor2.fetchall()
+        if(len(S_CID)>1):
+            command = f"SELECT * FROM `course` where "
+            for i in range(len(S_CID)):
+                if(i == len(S_CID)-1):
+                    command += f"CourseID = \"{S_CID[i][0]}\";"
+                else:
+                    command += f"CourseID = \"{S_CID[i][0]}\" or "
+            cursor2.execute(command)
+            data_a = cursor2.fetchall()
+
+        else:
+            command = f"SELECT * FROM `course` where CourseID = \"{S_CID[0][0]}\";"
+            cursor2.execute(command)
+            data_a = cursor2.fetchall()
+        data_b = opcourse(UID)
+        return data_a,data_b
+    else:
+        command = f"SELECT * FROM `course` ;"
+        cursor2.execute(command)
+        data_b = cursor2.fetchall()
+        return data_a,data_b
 
 
 @app.get('/')
@@ -213,54 +243,39 @@ def login(UID):
     if(flag):
         return "U"   
     else:
-        if(checkselected(UID)):
-            command = f"SELECT CourseID FROM `selectedcourse` where StudentID = \"{UID}\";"
-            cursor2.execute(command)
-            S_CID = cursor2.fetchall()
-            if(len(S_CID)>1):
-                command = f"SELECT * FROM `course` where "
-                for i in range(len(S_CID)):
-                    if(i == len(S_CID)-1):
-                        command += f"CourseID = \"{S_CID[i][0]}\";"
-                    else:
-                        command += f"CourseID = \"{S_CID[i][0]}\" or "
-                cursor2.execute(command)
-                data_a = cursor2.fetchall()
-    
-            else:
-                command = f"SELECT * FROM `course` where CourseID = \"{S_CID[0][0]}\";"
-                cursor2.execute(command)
-                data_a = cursor2.fetchall()
-            data_b = opcourse(UID)
-            return data_a,data_b
-        else:
-            command = f"SELECT * FROM `course` ;"
-            cursor2.execute(command)
-            data_b = cursor2.fetchall()
-            return 0,data_b
+        return data_return(UID)
 
 
-@app.get("/delcourse/{UID}/{CID}") #剔除不存在課號、學分數檢測
+@app.get("/delcourse/{UID}/{CID}") # 退選功能
 def delcourse(UID, CID):
     global data_a
-    if (checkCID(CID, data_a)):
-        if(checkcredit(CID, data_a)):
-            print("UID and CID",UID,CID)
-            command = f"DELETE FROM selectedcourse WHERE `selectedcourse`.`StudentID` = \"{UID}\" AND `selectedcourse`.`CourseID` = \"{CID}\";"
+    if(len(data_a)>0):
+        if (checkCID(CID, data_a)):  # 剔除不存在課號
+            if(checkcredit(CID, data_a)):  # 學分數檢測
+                command = f"DELETE FROM selectedcourse WHERE `selectedcourse`.`StudentID` = \"{UID}\" AND `selectedcourse`.`CourseID` = \"{CID}\";"
+                cursor2.execute(command)
+                c.commit()
+                return "S"
+            else:
+                return "N"
+        else:
+            return "F"
+    else:
+        return "D"
+
+@app.get("/addcourse/{UID}/{CID}")
+def addcourse(UID, CID):
+    global data_b
+    if(len(data_b)>0):
+        if (checkCID(CID, data_b)):
+            command = f"INSERT INTO `selectedcourse`(`StudentID`, `CourseID`) VALUES (\"{UID}\",\"{CID}\");"
             cursor2.execute(command)
-            print("command:",command)
-            print("s")
+            c.commit()
             return "S"
         else:
-            print("n")
-            return "N"
+            return "F"
     else:
-        print("f")
-        return "F"
+        return "D" 
 
-# @app.get("/addcourse/{UID}")
 
 c.commit()
-
-login(100001)
-delcourse(100001,1005)
